requirements.txt
aioredis==2.0.1
anyio==4.6.0
async-timeout==5.0.1
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
numpy==2.1.1
oauth2client==4.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
pytest==8.3.3
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
